requirements.txt
asgiref==3.5.2
coverage==6.4
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
